// MegabinFE/src/app/services/model/scheduledCollectionDto.ts
/**
 * Megabin API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ScheduledCollectionDto { 
    id: string;
    routeSequence: number;
    address: string | null;
    latitude: number;
    longitude: number;
    addressNotes: string | null;
    collected: boolean;
    notes: string | null;
    scheduledFor: string;
}